+++ src/sync.rs
@@ -1,4 +1,7 @@
-use std::collections::HashMap;
+use std::collections::{
+    HashMap,
+    HashSet,
+};
 
 use anyhow::Context;
 use futures::{
@@ -43,31 +46,33 @@ const CURSOR_VERSION: i64 = 1;
 /// returned to us every time Fivetran calls the `update` method of the
 /// connector.
 #[derive(Serialize, Deserialize, Debug, PartialEq, Eq, Clone)]
+#[serde(rename_all = "camelCase")]
 #[serde(deny_unknown_fields)]
 #[cfg_attr(test, derive(proptest_derive::Arbitrary))]
 pub struct State {
     /// The version of the connector that emitted this checkpoint. Could be used
     /// in the future to support backward compatibility with older state
     /// formats.
-    pub version: Option<i64>,
+    pub version: i64,
 
-    #[serde(default)]
     pub checkpoint: Checkpoint,
+
+    /// If set, then we are tracking the full set of tables that the connector
+    /// has every seen, so we are able to issue truncates the first time we
+    /// see a table.
+    ///
+    /// Older versions of state.json do not have this field set. Once all
+    /// state.json have this field, we can make this non-optional.
+    pub tables_seen: Option<HashSet<String>>,
 }
 
 impl State {
-    pub fn create(checkpoint: Checkpoint) -> Self {
+    pub fn create(checkpoint: Checkpoint, tables_seen: Option<HashSet<String>>) -> Self {
         Self {
-            version: Some(CURSOR_VERSION),
+            version: CURSOR_VERSION,
             checkpoint,
+            tables_seen,
         }
-    }
-}
-
-#[cfg(test)]
-impl Default for State {
-    fn default() -> Self {
-        Self::create(Checkpoint::default())
     }
 }
 
@@ -77,20 +82,11 @@ impl Default for State {
 pub enum Checkpoint {
     /// A checkpoint emitted during the initial synchonization.
     InitialSync {
-        snapshot: Option<i64>,
-        cursor: Option<ListSnapshotCursor>,
+        snapshot: i64,
+        cursor: ListSnapshotCursor,
     },
     /// A checkpoint emitted after an initial synchronzation has been completed.
     DeltaUpdates { cursor: DocumentDeltasCursor },
-}
-
-impl Default for Checkpoint {
-    fn default() -> Self {
-        Checkpoint::InitialSync {
-            snapshot: None,
-            cursor: None,
-        }
-    }
 }
 
 #[cfg(test)]
@@ -135,11 +131,12 @@ mod state_serialization_tests {
             )
             .unwrap(),
             State {
-                version: Some(1),
+                version: 1,
                 checkpoint: Checkpoint::InitialSync {
-                    snapshot: Some(42),
-                    cursor: Some(String::from("abc123").into()),
+                    snapshot: 42,
+                    cursor: String::from("abc123").into(),
                 },
+                tables_seen: None,
             },
         );
     }
@@ -152,24 +149,11 @@ mod state_serialization_tests {
             )
             .unwrap(),
             State {
-                version: Some(1),
+                version: 1,
                 checkpoint: Checkpoint::DeltaUpdates { cursor: 42.into() },
+                tables_seen: None,
             },
         );
-    }
-
-    #[test]
-    fn deserializes_initial_state() {
-        assert!(matches!(
-            serde_json::from_str::<State>("{}").unwrap(),
-            State {
-                version: None,
-                checkpoint: Checkpoint::InitialSync {
-                    snapshot: None,
-                    cursor: None
-                },
-            },
-        ));
     }
 }
 
@@ -234,13 +218,22 @@ impl From<UpdateMessage> for FivetranUpdateResponse {
 /// Returns the stream that the `update` endpoint emits.
 pub fn sync(
     source: impl Source + 'static,
-    state: State,
+    state: Option<State>,
 ) -> BoxStream<'static, anyhow::Result<UpdateMessage>> {
-    match state.checkpoint {
+    let Some(state) = state else {
+        return initial_sync(source, None, Some(HashSet::new())).boxed();
+    };
+
+    let State {
+        version: _version,
+        checkpoint,
+        tables_seen,
+    } = state;
+    match checkpoint {
         Checkpoint::InitialSync { snapshot, cursor } => {
-            initial_sync(source, snapshot, cursor).boxed()
+            initial_sync(source, Some((snapshot, cursor)), tables_seen).boxed()
         },
-        Checkpoint::DeltaUpdates { cursor } => delta_sync(source, cursor).boxed(),
+        Checkpoint::DeltaUpdates { cursor } => delta_sync(source, cursor, tables_seen).boxed(),
     }
 }
 
@@ -248,10 +241,10 @@ pub fn sync(
 #[try_stream(ok = UpdateMessage, error = anyhow::Error)]
 async fn initial_sync(
     source: impl Source,
-    snapshot: Option<i64>,
-    cursor: Option<ListSnapshotCursor>,
+    mut checkpoint: Option<(i64, ListSnapshotCursor)>,
+    mut tables_seen: Option<HashSet<String>>,
 ) {
-    let log_msg = if let Some(snapshot) = snapshot {
+    let log_msg = if let Some((snapshot, _)) = checkpoint {
         format!("Resuming an initial sync from {source} at {snapshot}")
     } else {
         format!("Starting an initial sync from {source}")
@@ -259,19 +252,27 @@ async fn initial_sync(
     log(&log_msg);
     yield UpdateMessage::Log(LogLevel::Info, log_msg);
 
-    let mut snapshot: Option<i64> = snapshot;
-    let mut cursor: Option<ListSnapshotCursor> = cursor;
     let mut has_more = true;
 
     while has_more {
-        yield UpdateMessage::Checkpoint(State::create(Checkpoint::InitialSync {
-            snapshot,
-            cursor: cursor.clone(),
-        }));
-
+        let snapshot = checkpoint.as_ref().map(|c| c.0);
+        let cursor = checkpoint.as_ref().map(|c| c.1.clone());
         let res = source.list_snapshot(snapshot, cursor.clone(), None).await?;
 
         for value in res.values {
+            if let Some(ref mut tables_seen) = tables_seen {
+                // Issue truncates if we see a table for the first time.
+                // Skip the behavior for legacy state.json - where tables_seen wasn't tracked.
+                if !tables_seen.contains(&value.table) {
+                    tables_seen.insert(value.table.clone());
+                    yield UpdateMessage::Update {
+                        schema_name: None,
+                        table_name: value.table.clone(),
+                        op_type: OpType::Truncate,
+                        row: HashMap::new(),
+                    };
+                }
+            }
             yield UpdateMessage::Update {
                 schema_name: None,
                 table_name: value.table,
@@ -281,12 +282,27 @@ async fn initial_sync(
         }
 
         has_more = res.has_more;
-        snapshot = Some(res.snapshot);
-        cursor = res.cursor.map(ListSnapshotCursor::from);
+        if has_more {
+            let cursor = ListSnapshotCursor::from(
+                res.cursor.context("Missing cursor when has_more was set")?,
+            );
+            yield UpdateMessage::Checkpoint(State::create(
+                Checkpoint::InitialSync {
+                    snapshot: res.snapshot,
+                    cursor: cursor.clone(),
+                },
+                tables_seen.clone(),
+            ));
+            checkpoint = Some((res.snapshot, cursor));
+        }
     }
 
-    let cursor = DocumentDeltasCursor::from(snapshot.context("Missing snapshot from response")?);
-    yield UpdateMessage::Checkpoint(State::create(Checkpoint::DeltaUpdates { cursor }));
+    let (snapshot, _) = checkpoint.context("list_snapshot lacking a snapshot for checkpoint")?;
+    let cursor = DocumentDeltasCursor::from(snapshot);
+    yield UpdateMessage::Checkpoint(State::create(
+        Checkpoint::DeltaUpdates { cursor },
+        tables_seen,
+    ));
 
     yield UpdateMessage::Log(LogLevel::Info, "Initial sync successful".to_string());
     log(&format!(
@@ -297,7 +313,11 @@ async fn initial_sync(
 /// Synchronizes the changes that happened after an initial synchronization or
 /// delta synchronization has been completed.
 #[try_stream(ok = UpdateMessage, error = anyhow::Error)]
-async fn delta_sync(source: impl Source, cursor: DocumentDeltasCursor) {
+async fn delta_sync(
+    source: impl Source,
+    cursor: DocumentDeltasCursor,
+    mut tables_seen: Option<HashSet<String>>,
+) {
     yield UpdateMessage::Log(
         LogLevel::Info,
         format!("Starting to apply changes from {source} starting at {cursor}"),
@@ -310,6 +330,20 @@ async fn delta_sync(source: impl Source, cursor: DocumentDeltasCursor) {
         let response = source.document_deltas(cursor, None).await?;
 
         for value in response.values {
+            if let Some(ref mut tables_seen) = tables_seen {
+                // Issue truncates if we see a table for the first time.
+                // Skip the behavior for legacy state.json - where tables_seen wasn't tracked.
+                if !tables_seen.contains(&value.table) {
+                    tables_seen.insert(value.table.clone());
+                    yield UpdateMessage::Update {
+                        schema_name: None,
+                        table_name: value.table.clone(),
+                        op_type: OpType::Truncate,
+                        row: HashMap::new(),
+                    };
+                }
+            }
+
             yield UpdateMessage::Update {
                 schema_name: None,
                 table_name: value.table,
@@ -327,7 +361,10 @@ async fn delta_sync(source: impl Source, cursor: DocumentDeltasCursor) {
 
         // It is safe to take a snapshot here, because document_deltas
         // guarantees that the state given by one call is consistent.
-        yield UpdateMessage::Checkpoint(State::create(Checkpoint::DeltaUpdates { cursor }));
+        yield UpdateMessage::Checkpoint(State::create(
+            Checkpoint::DeltaUpdates { cursor },
+            tables_seen.clone(),
+        ));
     }
 
     yield UpdateMessage::Log(LogLevel::Info, "Changes applied".to_string());